Source/Backend/requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudinary==1.42.1
cryptography==43.0.3
dj-database-url==2.3.0
Django==5.1.2
django-allauth==65.2.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-oauth-toolkit==3.0.1
django-picklefield==3.2
django-q2==1.7.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
jinxed==1.3.0
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.2
psycopg==3.2.3
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==3.5.3
requests==2.32.3
setuptools==75.5.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
stream-chat==5.11.0
wcwidth==0.2.13
